// src/lib/api/models/getApiFieldGroupsEntityType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiFieldGroupsEntityType = typeof GetApiFieldGroupsEntityType[keyof typeof GetApiFieldGroupsEntityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiFieldGroupsEntityType = {
  PROPERTY: 'PROPERTY',
  CUSTOMER_NEED: 'CUSTOMER_NEED',
  LEAD: 'LEAD',
  DEAL: 'DEAL',
  OWNER_PROFILE: 'OWNER_PROFILE',
} as const;