// src/lib/api/models/getApiFormSchemasEntityType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiFormSchemasEntityType = typeof GetApiFormSchemasEntityType[keyof typeof GetApiFormSchemasEntityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiFormSchemasEntityType = {
  PROPERTY: 'PROPERTY',
  CUSTOMER_NEED: 'CUSTOMER_NEED',
  LEAD: 'LEAD',
  DEAL: 'DEAL',
} as const;